Rename `function` module to `func`.
