+++ src/utils/array.js
@@ -1,5 +1,5 @@
 import DateUtil from './date.js';
-import FunctionUtil from './function.js';
+import FuncUtil from './func.js';
 import MathUtil from './math.js';
 import ObjectUtil from './object.js';
 import RandomUtil from './random.js';
